+++ business-easy-core/src/main/java/com/businesseasy/core/controllers/PeopleController.java
@@ -32,7 +32,7 @@ public class PeopleController {
     }
 
     @PostMapping("")
-    Integer insertPeople(@Valid @RequestBody People request) {
+    Long insertPeople(@Valid @RequestBody People request) {
         return peopleService.insertPeople(request);
     }
