+++ business-easy-core/src/main/java/com/businesseasy/core/services/people/PeopleService.java
@@ -12,7 +12,7 @@ import java.util.Map;
 public interface PeopleService {
     Page<PeopleEntity> getPeople(Map<String, String> parameterMap);
 
-    Integer insertPeople(People request);
+    Long insertPeople(People request);
 
     People updatePeople(People request);
